% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eavsr-package.R
\docType{package}
\name{eavsr-package}
\alias{eavsr}
\alias{eavsr-package}
\title{eavsr: Tidy Access to the Election Administration and Voting Survey}
\description{
Downloads, caches, and harmonizes public-release files from the U.S.
Election Assistance Commission's Election Administration and Voting
Survey (EAVS). Provides tidy access to multiple survey cycles, exposes
the codebook as a tibble, and supports joins to county-level rurality
classifications via the \code{rurality} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwimpy/eavsr}
  \item \url{https://cwimpy.github.io/eavsr/}
  \item Report bugs at \url{https://github.com/cwimpy/eavsr/issues}
}

}
\author{
\strong{Maintainer}: Cameron Wimpy \email{cwimpy@astate.edu} (\href{https://orcid.org/0000-0002-2049-5229}{ORCID})

Authors:
\itemize{
  \item Cameron Wimpy \email{cwimpy@astate.edu} (\href{https://orcid.org/0000-0002-2049-5229}{ORCID})
}

}
\keyword{internal}
